requirements.txt
discord
pyyaml
matplotlib